refactor defi-basket
