--- projects/defi-basket/abi.json
@@ -1,7 +1,0 @@
-{
-  "inputs": [],
-  "name": "tokenCounter",
-  "outputs": [{ "internalType": "uint256", "name": "", "type": "uint256" }],
-  "stateMutability": "view",
-  "type": "function"
-}
